// supabase/types.d.ts
export interface TaskRow {
  id: number;
  description: string;
  due_date: string | null;
  completed: boolean;
  created_at: string;
  completed_at: string | null;
  user_id: string;
  guild_id: string;
}
export interface TaskInsert {
  description: string;
  due_date?: string | null;
  completed?: boolean;
  created_at?: string;
  completed_at?: string | null;
  user_id: string;
  guild_id: string;
}
export interface TaskUpdate {
  description?: string;
  due_date?: string | null;
  completed?: boolean;
  completed_at?: string | null;
}